devise authentication and answer creation features added

+++ app/controllers/questions_controller.rb
@@ -1,10 +1,12 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %w(index show)
   expose :questions, ->{ Question.all }
-  expose :question
+  expose :question, build: -> (question_params){ current_user.questions.new(question_params) }
+  expose :answer, ->{question.answers.new}
 
   def create
     if question.save
-      redirect_to question
+      redirect_to question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -19,8 +21,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question succesfully deleted.'
+    else
+      redirect_to question, alert: 'You are not the author of this question.'
+    end
   end
 
   private
